+++ packages/core/components/checkbox/index.js
@@ -4,6 +4,7 @@ import styled, { css } from "styled-components";
 import { themeGet } from "styled-system";
 import Icon from "../icon";
 import { COMMON, FLEX_ITEM, LAYOUT, MISC, POSITION } from "../../constants";
+import { blurUnlessFocusVisible } from "../../utils";
 
 const StyledCheckbox = styled.span`
   position: relative;
@@ -116,6 +117,7 @@ const Checkbox = ({
         name={name}
         value={value}
         onChange={e => onChange(e, e.target.value)}
+        onFocus={e => blurUnlessFocusVisible(e.target)}
         disabled={disabled}
         defaultChecked={defaultChecked}
         checked={indeterminate || checked}