+++ packages/core/theme/components/focusVisible/index.js
@@ -1,0 +1,5 @@
+import { css } from "styled-components";
+
+export const theme = {
+    styles: css``
+};